Enhance GitHub showcase exhibit
In this commit, I've refined the README to create a more engaging and informative GitHub showcase. The introduction now features a friendly greeting with emoji flair, emphasizing the purpose of the exhibit to highlight projects mentioned in my public CV. Special attention is drawn to the 'timepeace' project, showcasing its organized wiki and vital links. Additionally, privacy considerations are addressed by adhering to ICO standards and ensuring the removal or alteration of identifiable information and potential risks like API keys. This commit aims to create a professional yet warm and inviting online presence.